requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart

# Computer Vision
# IMPORTANT: Keep numpy below 2.0 to avoid breaking BasicSR/Real-ESRGAN
numpy<2.0.0  
pillow
opencv-python-headless
imageio

# AI & ML
# Relaxed versions to fit your Codespace Python version
torch>=2.2.0
torchvision>=0.17.0

# Face & Upscaling Logic
gfpgan>=1.3.8
realesrgan>=0.3.0
basicsr>=1.4.2
mediapipe>=0.10.9

# Cloud Storage
google-cloud-storage

# Utilities
requests
tqdm